feat: remove assets when uploading new

+++ controllers/admin/AdminCoursesController.js
@@ -73,6 +73,7 @@ module.exports.editCourse = async function(req, res) {
   }
 };
 module.exports.createCourse = async function(req, res) {
+  const curriculumPdf = req.body.curriculumPdf ? `${req.body.curriculumPdf.split('.')[0]}_${uuid(4)}.${req.body.curriculumPdf.split('.').reverse()[0]}` : undefined
   const storys = await Story.find()
     .select("title slug")
     .exec();
@@ -100,6 +101,7 @@ module.exports.createCourse = async function(req, res) {
   course.requiredLanguages = req.body.requiredLanguages;
   course.joinTheTechDisruption = req.body.joinTheTechDisruption;
   course.startYourClass = req.body.startYourClass;
+  course.curriculumPdf = req.body.curriculumPdf ? req.body.curriculumPdf : undefined;
 
   if(!!req.body.successStory) {
     course.successStory = req.body.successStory;
@@ -129,7 +131,6 @@ module.exports.createCourse = async function(req, res) {
     };
   });
 
-  course.curriculumPdf = req.body.curriculumPdf;
 
   // save the course and check for errors
   course.save(async function(err) {
@@ -220,10 +221,10 @@ exports.resizeImages = async (request, response, next) => {
     try {
       if (singleFile[0].mimetype === "image/svg+xml" || singleFile[0].mimetype === "application/pdf") {
         const fileObject = fs.readFileSync(singleFile[0].path);
+        const filename = `${request.body[singleFile[0].fieldname].split('.')[0]}_${uuid(4)}.${request.body[singleFile[0].fieldname].split('.').reverse()[0]}`
+        request.body[singleFile[0].fieldname] = filename
         fs.writeFileSync(
-          `${process.env.IMAGE_UPLOAD_DIR}/${
-            request.body[singleFile[0].fieldname]
-          }`,
+          `${process.env.IMAGE_UPLOAD_DIR}${ filename }`,
           fileObject
         );
       }
@@ -235,7 +236,9 @@ exports.resizeImages = async (request, response, next) => {
             request.body[singleFile[0].fieldname]
           }`
         );
-        fs.unlinkSync(singleFile[0].path);
+        if(fs.existsSync(singleFile[0].path)){
+          fs.unlinkSync(singleFile[0].path);
+        }
       }
     } catch (error) {
       console.log(error);
@@ -246,6 +249,53 @@ exports.resizeImages = async (request, response, next) => {
 
 module.exports.updateCourse = async function(req, res) {
   let course = await Course.findOne({ slug: req.params.slug });
+
+  if(req.body.curriculumPdf && await fs.existsSync(`${process.env.IMAGE_UPLOAD_DIR}${course.curriculumPdf}`)) {
+    await fs.unlinkSync(`${process.env.IMAGE_UPLOAD_DIR}${course.curriculumPdf}`);
+  }
+  if(req.body.icon && await fs.existsSync(`${process.env.IMAGE_UPLOAD_DIR}${course.icon}`)) {
+    await fs.unlinkSync(`${process.env.IMAGE_UPLOAD_DIR}${course.icon}`);
+  }
+  if(req.body.subicon && await fs.existsSync(`${process.env.IMAGE_UPLOAD_DIR}${course.subicon}`)) {
+    await fs.unlinkSync(`${process.env.IMAGE_UPLOAD_DIR}${course.subicon}`);
+  }
+  if(req.body.archivement_icon_1 && await fs.existsSync(`${process.env.IMAGE_UPLOAD_DIR}${course.archivement_icon_1}`)) {
+    await fs.unlinkSync(`${process.env.IMAGE_UPLOAD_DIR}${course.archivement_icon_1}`);
+  }
+  if(req.body.archivement_icon_2 && await fs.existsSync(`${process.env.IMAGE_UPLOAD_DIR}${course.archivement_icon_2}`)) {
+    await fs.unlinkSync(`${process.env.IMAGE_UPLOAD_DIR}${course.archivement_icon_2}`);
+  }
+  if(req.body.archivement_icon_3 && await fs.existsSync(`${process.env.IMAGE_UPLOAD_DIR}${course.archivement_icon_3}`)) {
+    await fs.unlinkSync(`${process.env.IMAGE_UPLOAD_DIR}${course.archivement_icon_3}`);
+  }
+  if(req.body.archivement_icon_4 && await fs.existsSync(`${process.env.IMAGE_UPLOAD_DIR}${course.archivement_icon_4}`)) {
+    await fs.unlinkSync(`${process.env.IMAGE_UPLOAD_DIR}${course.archivement_icon_4}`);
+  }
+  if(req.body.archivement_icon_5 && await fs.existsSync(`${process.env.IMAGE_UPLOAD_DIR}${course.archivement_icon_5}`)) {
+    await fs.unlinkSync(`${process.env.IMAGE_UPLOAD_DIR}${course.archivement_icon_5}`);
+  }
+  if(req.body.archivement_icon_6 && await fs.existsSync(`${process.env.IMAGE_UPLOAD_DIR}${course.archivement_icon_6}`)) {
+    await fs.unlinkSync(`${process.env.IMAGE_UPLOAD_DIR}${course.archivement_icon_6}`);
+  }
+  if(req.body.features_icon_1 && await fs.existsSync(`${process.env.IMAGE_UPLOAD_DIR}${course.features_icon_1}`)) {
+    await fs.unlinkSync(`${process.env.IMAGE_UPLOAD_DIR}${course.features_icon_1}`);
+  }
+  if(req.body.features_icon_2 && await fs.existsSync(`${process.env.IMAGE_UPLOAD_DIR}${course.features_icon_2}`)) {
+    await fs.unlinkSync(`${process.env.IMAGE_UPLOAD_DIR}${course.features_icon_2}`);
+  }
+  if(req.body.features_icon_3 && await fs.existsSync(`${process.env.IMAGE_UPLOAD_DIR}${course.features_icon_3}`)) {
+    await fs.unlinkSync(`${process.env.IMAGE_UPLOAD_DIR}${course.features_icon_3}`);
+  }
+  if(req.body.features_icon_4 && await fs.existsSync(`${process.env.IMAGE_UPLOAD_DIR}${course.features_icon_4}`)) {
+    await fs.unlinkSync(`${process.env.IMAGE_UPLOAD_DIR}${course.features_icon_4}`);
+  }
+  if(req.body.features_icon_5 && await fs.existsSync(`${process.env.IMAGE_UPLOAD_DIR}${course.features_icon_5}`)) {
+    await fs.unlinkSync(`${process.env.IMAGE_UPLOAD_DIR}${course.features_icon_5}`);
+  }
+  if(req.body.features_icon_6 && await fs.existsSync(`${process.env.IMAGE_UPLOAD_DIR}${course.features_icon_6}`)) {
+    await fs.unlinkSync(`${process.env.IMAGE_UPLOAD_DIR}${course.features_icon_6}`);
+  }
+  
   course.slug = req.body.slug;
   course.icon = req.body.icon ? req.body.icon : course.icon;
   course.coloraccent = req.body.coloraccent ? req.body.coloraccent : "";
@@ -258,9 +308,8 @@ module.exports.updateCourse = async function(req, res) {
   course.massnahmeNumber = req.body.massnahmenummer;
   course.massnahmeDetails = req.body.massnahmedetails;
   course.courselength = req.body.courselength;
-  course.curriculumPdf = req.body.curriculumPdf;
   course.locations = req.body.locations;
-  course.curriculumPdf = req.body.curriculumPdf;
+  course.curriculumPdf = req.body.curriculumPdf ? req.body.curriculumPdf : course.curriculumPdf;
   course.icon = req.files.icon ? req.body.icon : course.icon;
   course.subicon = req.files.subicon ? req.body.subicon : course.subicon;
   course.curriculumSectionSubheading = req.body.curriculumSectionSubheading;
